requirements.txt
Django==1.6
South==0.8.4
argparse==1.2.1
boto==2.19.0
django-crispy-forms==1.4.0
oauthlib==0.6.0
python-openid==2.2.5
python-social-auth==0.1.17
requests==2.0.1
requests-oauthlib==0.4.0
six==1.4.1
wsgiref==0.1.2
